+++ ca/django_ca/models.py
@@ -541,7 +541,13 @@ class X509CertMixin(DjangoCAModel):
     @cached_property
     def authority_information_access(self) -> Optional[AuthorityInformationAccess]:  # pragma: no cover
         """The ``django_ca.extensions.AuthorityInformationAccess`` extension or ``None`` if not
-        present."""
+        present.
+
+        .. deprecated:: 1.22.0
+
+           Extension wrapper classes are deprecated and will be removed in ``django-ca==1.24.0``. Use
+           :py:attr:`~django_ca.models.X509CertMixin.x509_extensions` instead.
+        """
         try:
             ext = self.pub.loaded.extensions.get_extension_for_class(x509.AuthorityInformationAccess)
         except x509.ExtensionNotFound:
@@ -551,7 +557,13 @@ class X509CertMixin(DjangoCAModel):
     @cached_property
     def authority_key_identifier(self) -> Optional[AuthorityKeyIdentifier]:  # pragma: no cover
         """The ``django_ca.extensions.AuthorityKeyIdentifier`` extension or ``None`` if not
-        present."""
+        present.
+
+        .. deprecated:: 1.22.0
+
+           Extension wrapper classes are deprecated and will be removed in ``django-ca==1.24.0``. Use
+           :py:attr:`~django_ca.models.X509CertMixin.x509_extensions` instead.
+        """
         try:
             ext = self.pub.loaded.extensions.get_extension_for_class(x509.AuthorityKeyIdentifier)
         except x509.ExtensionNotFound:
@@ -560,7 +572,13 @@ class X509CertMixin(DjangoCAModel):
 
     @cached_property
     def basic_constraints(self) -> Optional[BasicConstraints]:  # pragma: no cover
-        """The ``django_ca.extensions.BasicConstraints`` extension or ``None`` if not present."""
+        """The ``django_ca.extensions.BasicConstraints`` extension or ``None`` if not present.
+
+        .. deprecated:: 1.22.0
+
+           Extension wrapper classes are deprecated and will be removed in ``django-ca==1.24.0``. Use
+           :py:attr:`~django_ca.models.X509CertMixin.x509_extensions` instead.
+        """
         try:
             ext = self.pub.loaded.extensions.get_extension_for_class(x509.BasicConstraints)
         except x509.ExtensionNotFound:
@@ -570,7 +588,13 @@ class X509CertMixin(DjangoCAModel):
     @cached_property
     def crl_distribution_points(self) -> Optional[CRLDistributionPoints]:  # pragma: no cover
         """The ``django_ca.extensions.CRLDistributionPoints`` extension or ``None`` if not
-        present."""
+        present.
+
+        .. deprecated:: 1.22.0
+
+           Extension wrapper classes are deprecated and will be removed in ``django-ca==1.24.0``. Use
+           :py:attr:`~django_ca.models.X509CertMixin.x509_extensions` instead.
+        """
         try:
             ext = self.pub.loaded.extensions.get_extension_for_class(x509.CRLDistributionPoints)
         except x509.ExtensionNotFound:
@@ -579,7 +603,13 @@ class X509CertMixin(DjangoCAModel):
 
     @cached_property
     def certificate_policies(self) -> Optional[CertificatePolicies]:  # pragma: no cover
-        """The ``django_ca.extensions.CertificatePolicies`` extension or ``None`` if not present."""
+        """The ``django_ca.extensions.CertificatePolicies`` extension or ``None`` if not present.
+
+        .. deprecated:: 1.22.0
+
+           Extension wrapper classes are deprecated and will be removed in ``django-ca==1.24.0``. Use
+           :py:attr:`~django_ca.models.X509CertMixin.x509_extensions` instead.
+        """
         try:
             ext = self.pub.loaded.extensions.get_extension_for_class(x509.CertificatePolicies)
         except x509.ExtensionNotFound:
@@ -588,7 +618,13 @@ class X509CertMixin(DjangoCAModel):
 
     @cached_property
     def freshest_crl(self) -> Optional[FreshestCRL]:  # pragma: no cover
-        """The ``django_ca.extensions.FreshestCRL`` extension or ``None`` if not present."""
+        """The ``django_ca.extensions.FreshestCRL`` extension or ``None`` if not present.
+
+        .. deprecated:: 1.22.0
+
+           Extension wrapper classes are deprecated and will be removed in ``django-ca==1.24.0``. Use
+           :py:attr:`~django_ca.models.X509CertMixin.x509_extensions` instead.
+        """
         try:
             ext = self.pub.loaded.extensions.get_extension_for_class(x509.FreshestCRL)
         except x509.ExtensionNotFound:
@@ -597,7 +633,13 @@ class X509CertMixin(DjangoCAModel):
 
     @cached_property
     def inhibit_any_policy(self) -> Optional[InhibitAnyPolicy]:  # pragma: no cover
-        """The ``django_ca.extensions.InhibitAnyPolicy`` extension or ``None`` if not present."""
+        """The ``django_ca.extensions.InhibitAnyPolicy`` extension or ``None`` if not present.
+
+        .. deprecated:: 1.22.0
+
+           Extension wrapper classes are deprecated and will be removed in ``django-ca==1.24.0``. Use
+           :py:attr:`~django_ca.models.X509CertMixin.x509_extensions` instead.
+        """
         try:
             ext = self.pub.loaded.extensions.get_extension_for_class(x509.InhibitAnyPolicy)
         except x509.ExtensionNotFound:
@@ -607,7 +649,13 @@ class X509CertMixin(DjangoCAModel):
     @cached_property
     def issuer_alternative_name(self) -> Optional[IssuerAlternativeName]:  # pragma: no cover
         """The ``django_ca.extensions.IssuerAlternativeName`` extension or ``None`` if not
-        present."""
+        present.
+
+        .. deprecated:: 1.22.0
+
+           Extension wrapper classes are deprecated and will be removed in ``django-ca==1.24.0``. Use
+           :py:attr:`~django_ca.models.X509CertMixin.x509_extensions` instead.
+        """
         try:
             ext = self.pub.loaded.extensions.get_extension_for_class(x509.IssuerAlternativeName)
         except x509.ExtensionNotFound:
@@ -616,7 +664,13 @@ class X509CertMixin(DjangoCAModel):
 
     @cached_property
     def policy_constraints(self) -> Optional[PolicyConstraints]:  # pragma: no cover
-        """The ``django_ca.extensions.PolicyConstraints`` extension or ``None`` if not present."""
+        """The ``django_ca.extensions.PolicyConstraints`` extension or ``None`` if not present.
+
+        .. deprecated:: 1.22.0
+
+           Extension wrapper classes are deprecated and will be removed in ``django-ca==1.24.0``. Use
+           :py:attr:`~django_ca.models.X509CertMixin.x509_extensions` instead.
+        """
         try:
             ext = self.pub.loaded.extensions.get_extension_for_class(x509.PolicyConstraints)
         except x509.ExtensionNotFound:
@@ -625,7 +679,13 @@ class X509CertMixin(DjangoCAModel):
 
     @cached_property
     def key_usage(self) -> Optional[KeyUsage]:  # pragma: no cover
-        """The ``django_ca.extensions.KeyUsage`` extension or ``None`` if not present."""
+        """The ``django_ca.extensions.KeyUsage`` extension or ``None`` if not present.
+
+        .. deprecated:: 1.22.0
+
+           Extension wrapper classes are deprecated and will be removed in ``django-ca==1.24.0``. Use
+           :py:attr:`~django_ca.models.X509CertMixin.x509_extensions` instead.
+        """
         try:
             ext = self.pub.loaded.extensions.get_extension_for_class(x509.KeyUsage)
         except x509.ExtensionNotFound:
@@ -634,7 +694,13 @@ class X509CertMixin(DjangoCAModel):
 
     @cached_property
     def extended_key_usage(self) -> Optional[ExtendedKeyUsage]:  # pragma: no cover
-        """The ``django_ca.extensions.ExtendedKeyUsage`` extension or ``None`` if not present."""
+        """The ``django_ca.extensions.ExtendedKeyUsage`` extension or ``None`` if not present.
+
+        .. deprecated:: 1.22.0
+
+           Extension wrapper classes are deprecated and will be removed in ``django-ca==1.24.0``. Use
+           :py:attr:`~django_ca.models.X509CertMixin.x509_extensions` instead.
+        """
         try:
             ext = self.pub.loaded.extensions.get_extension_for_class(x509.ExtendedKeyUsage)
         except x509.ExtensionNotFound:
@@ -643,7 +709,13 @@ class X509CertMixin(DjangoCAModel):
 
     @cached_property
     def name_constraints(self) -> Optional[NameConstraints]:  # pragma: no cover
-        """The ``django_ca.extensions.NameConstraints`` extension or ``None`` if not present."""
+        """The ``django_ca.extensions.NameConstraints`` extension or ``None`` if not present.
+
+        .. deprecated:: 1.22.0
+
+           Extension wrapper classes are deprecated and will be removed in ``django-ca==1.24.0``. Use
+           :py:attr:`~django_ca.models.X509CertMixin.x509_extensions` instead.
+        """
         try:
             ext = self.pub.loaded.extensions.get_extension_for_class(x509.NameConstraints)
         except x509.ExtensionNotFound:
@@ -652,7 +724,13 @@ class X509CertMixin(DjangoCAModel):
 
     @cached_property
     def ocsp_no_check(self) -> Optional[OCSPNoCheck]:  # pragma: no cover
-        """The ``django_ca.extensions.OCSPNoCheck`` extension or ``None`` if not present."""
+        """The ``django_ca.extensions.OCSPNoCheck`` extension or ``None`` if not present.
+
+        .. deprecated:: 1.22.0
+
+           Extension wrapper classes are deprecated and will be removed in ``django-ca==1.24.0``. Use
+           :py:attr:`~django_ca.models.X509CertMixin.x509_extensions` instead.
+        """
         try:
             ext = self.pub.loaded.extensions.get_extension_for_class(x509.OCSPNoCheck)
         except x509.ExtensionNotFound:
@@ -661,7 +739,13 @@ class X509CertMixin(DjangoCAModel):
 
     @cached_property
     def precert_poison(self) -> Optional[PrecertPoison]:  # pragma: no cover
-        """The ``django_ca.extensions.PrecertPoison`` extension or ``None`` if not present."""
+        """The ``django_ca.extensions.PrecertPoison`` extension or ``None`` if not present.
+
+        .. deprecated:: 1.22.0
+
+           Extension wrapper classes are deprecated and will be removed in ``django-ca==1.24.0``. Use
+           :py:attr:`~django_ca.models.X509CertMixin.x509_extensions` instead.
+        """
         try:
             self.pub.loaded.extensions.get_extension_for_class(x509.PrecertPoison)
         except x509.ExtensionNotFound:
@@ -673,7 +757,13 @@ class X509CertMixin(DjangoCAModel):
         self,
     ) -> Optional[PrecertificateSignedCertificateTimestamps]:  # pragma: no cover
         """The ``django_ca.extensions.PrecertificateSignedCertificateTimestamps`` extension or
-        ``None`` if not present."""
+        ``None`` if not present.
+
+        .. deprecated:: 1.22.0
+
+           Extension wrapper classes are deprecated and will be removed in ``django-ca==1.24.0``. Use
+           :py:attr:`~django_ca.models.X509CertMixin.x509_extensions` instead.
+        """
         try:
             ext = self.pub.loaded.extensions.get_extension_for_class(
                 x509.PrecertificateSignedCertificateTimestamps
@@ -685,7 +775,13 @@ class X509CertMixin(DjangoCAModel):
     @cached_property
     def subject_alternative_name(self) -> Optional[SubjectAlternativeName]:  # pragma: no cover
         """The ``django_ca.extensions.SubjectAlternativeName`` extension or ``None`` if not
-        present."""
+        present.
+
+        .. deprecated:: 1.22.0
+
+           Extension wrapper classes are deprecated and will be removed in ``django-ca==1.24.0``. Use
+           :py:attr:`~django_ca.models.X509CertMixin.x509_extensions` instead.
+        """
         try:
             ext = self.pub.loaded.extensions.get_extension_for_class(x509.SubjectAlternativeName)
         except x509.ExtensionNotFound:
@@ -694,7 +790,13 @@ class X509CertMixin(DjangoCAModel):
 
     @cached_property
     def subject_key_identifier(self) -> Optional[SubjectKeyIdentifier]:  # pragma: no cover
-        """The ``django_ca.extensions.SubjectKeyIdentifier`` extension or ``None`` if not present."""
+        """The ``django_ca.extensions.SubjectKeyIdentifier`` extension or ``None`` if not present.
+
+        .. deprecated:: 1.22.0
+
+           Extension wrapper classes are deprecated and will be removed in ``django-ca==1.24.0``. Use
+           :py:attr:`~django_ca.models.X509CertMixin.x509_extensions` instead.
+        """
         try:
             ext = self.pub.loaded.extensions.get_extension_for_class(x509.SubjectKeyIdentifier)
         except x509.ExtensionNotFound:
@@ -703,7 +805,13 @@ class X509CertMixin(DjangoCAModel):
 
     @cached_property
     def tls_feature(self) -> Optional[TLSFeature]:  # pragma: no cover
-        """The ``django_ca.extensions.TLSFeature`` extension or ``None`` if not present."""
+        """The ``django_ca.extensions.TLSFeature`` extension or ``None`` if not present.
+
+        .. deprecated:: 1.22.0
+
+           Extension wrapper classes are deprecated and will be removed in ``django-ca==1.24.0``. Use
+           :py:attr:`~django_ca.models.X509CertMixin.x509_extensions` instead.
+        """
         try:
             ext = self.pub.loaded.extensions.get_extension_for_class(x509.TLSFeature)
         except x509.ExtensionNotFound: